refactor: 변수 타입 및 애노테이션 수정

+++ src/main/java/com/allclearwas/domains/enrollment/dto/response/CourseEnrollmentCountRes.java
@@ -4,7 +4,7 @@ import com.allclearwas.domains.enrollment.dao.CourseEnrollmentCountDao;
 
 public record CourseEnrollmentCountRes(
 	Long courseId,
-	int current
+	Long current
 ) {
 	public static CourseEnrollmentCountRes of(CourseEnrollmentCountDao dao) {
 		return new CourseEnrollmentCountRes(dao.getCourseId(), dao.getCount());
